剔除CQ

+++ models/config.go
@@ -37,8 +37,8 @@ type Yaml struct {
 	Theme               string
 	TelegramBotToken    string `yaml:"telegram_bot_token"`
 	TelegramUserID      int    `yaml:"telegram_user_id"`
-	QQID                int64  `yaml:"qquid"`
-	QQGroupID           int64  `yaml:"qqgid"`
+	QQID                int    `yaml:"qquid"`
+	QQGroupID           int    `yaml:"qqgid"`
 	DefaultPriority     int    `yaml:"default_priority"`
 	NoGhproxy           bool   `yaml:"no_ghproxy"`
 	QbotPublicMode      bool   `yaml:"qbot_public_mode"`
